+++ src/execution/vm/bytecode_handlers.cpp
@@ -1,11 +1,11 @@
 #include "execution/vm/bytecode_handlers.h"
 
-#include "self_driving/brain_defs.h"
 #include "catalog/catalog_defs.h"
 #include "execution/exec/execution_context.h"
 #include "execution/sql/index_iterator.h"
 #include "execution/sql/storage_interface.h"
 #include "execution/sql/vector_projection_iterator.h"
+#include "self_driving/operating_unit_defs.h"
 
 extern "C" {
