+++ src/include/self_driving/brain_util.h
@@ -2,20 +2,13 @@
 
 #include <string>
 
-#include "self_driving/brain_defs.h"
+#include "self_driving/operating_unit_defs.h"
 
 namespace noisepage::selfdriving {
 
 /**
  * Utility class for helper functions
  */
-class BrainUtil {
- public:
-  /**
-   * Converts an ExecutionOperatingUnitType enum to string representation
-   * @param f ExecutionOperatingUnitType to convert
-   */
-  static std::string ExecutionOperatingUnitTypeToString(ExecutionOperatingUnitType f);
-};
+class BrainUtil {};
 
 }  // namespace noisepage::selfdriving
