+++ src/include/self_driving/operating_unit.h
@@ -5,10 +5,10 @@
 #include <utility>
 #include <vector>
 
-#include "self_driving/brain_defs.h"
 #include "execution/exec_defs.h"
 #include "execution/sql/memory_pool.h"
 #include "execution/util/execution_common.h"
+#include "self_driving/operating_unit_defs.h"
 
 namespace noisepage::runner {
 class MiniRunners;
