+++ util/runner/mini_runners_argument_generator.cpp
@@ -1,8 +1,9 @@
+#include "runner/mini_runners_argument_generator.h"
+
 #include <cmath>
 #include <unordered_set>
 
-#include "self_driving/brain_defs.h"
-#include "runner/mini_runners_argument_generator.h"
+#include "self_driving/operating_unit_defs.h"
 
 namespace noisepage::runner {
